examples/llm_compression/torch/qat_with_lora/requirements.txt
tqdm
whowhatbench @ git+https://github.com/openvinotoolkit/openvino.genai#subdirectory=tools/who_what_benchmark
numpy>=1.23.5,<2
openvino==2025.0
optimum-intel>=1.22.0
transformers>=4.48.0